Windows: Avoid C4267.

+++ impl/weekly.c
@@ -129,8 +129,8 @@ status week_init(a_week *week, const char *s, size_t len)
 void week_add_to_schedule(const a_week *week, const a_time *t, a_schedule *schedule)
 {
   a_time date_ = time_clone(t), *date = &date_;
-  size_t i = 0;
-  for (; i < DIM(week->days); ++i) {
+  int i = 0;
+  for (; i < (int)DIM(week->days); ++i) {
     const a_day *day = &week->days[i];
     if (0 == day->n_ranges) continue;
     time_whms(date, i, 0, 0, 0);
